requirements.txt
pytest-xdist
pytest-ordering
pytest-rerunfailures
pytest-html
allure-pytest
pyyaml
requests